Adds emplyee validaton and fixes department change

+++ DepartmentManagementAsp/Controllers/EmployeeController.cs
@@ -1,9 +1,11 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using DepartmentManagementModels;
 using DepartmentManagementModels.OperationResults;
 using DepartmentManagementModels.Repositories;
+using DepartmentManagementModels.Validators;
 using Microsoft.AspNetCore.Mvc;
 
 namespace DepartmentManagementAsp.Controllers
@@ -12,7 +14,13 @@ namespace DepartmentManagementAsp.Controllers
     public class EmployeeController : Controller
     {
         private IEmployeeRepository EmployeeRepository { get; }
-        public EmployeeController(IEmployeeRepository employeeRepository) => EmployeeRepository = employeeRepository;
+        private IEmployeeValidator EmployeeValidator { get; }
+
+        public EmployeeController(IEmployeeRepository employeeRepository, IEmployeeValidator employeeValidator)
+        {
+            EmployeeRepository = employeeRepository;
+            EmployeeValidator = employeeValidator;
+        }
 
         [HttpGet]
         public IActionResult GetEmployees([FromQuery] long departmentId = 0, bool includeDepartment = false)
@@ -21,13 +29,17 @@ namespace DepartmentManagementAsp.Controllers
                 ? EmployeeRepository.GetEmployees(includeDepartment)
                 : EmployeeRepository.GetEmployees(includeDepartment, e => e.Department.DepartmentId == departmentId);
             
-            
             return Ok(BreakDepReferenceCycle(employees));
         }
 
         [HttpPost]
         public async Task<IActionResult> PostEmployee([FromBody] Employee employee)
         {
+            if (ValidateEmployeeOrBadRequest(employee, EmployeeValidator.ValidateOnAdd) is BadRequestObjectResult badRequest)
+            {
+                return badRequest;
+            }
+
             OperationResult operationResult = await EmployeeRepository.AddEmployeeAsync(employee);
             return GetResultFromOperationResult(operationResult);
         }
@@ -35,6 +47,11 @@ namespace DepartmentManagementAsp.Controllers
         [HttpPut]
         public async Task<IActionResult> PutEmployee([FromBody] Employee employee)
         {
+            if (ValidateEmployeeOrBadRequest(employee, EmployeeValidator.ValidateOnEdit) is BadRequestObjectResult badRequest)
+            {
+                return badRequest;
+            }
+
             OperationResult operationResult = await EmployeeRepository.EditEmployeeAsync(employee);
             return GetResultFromOperationResult(operationResult);
         }
@@ -42,14 +59,25 @@ namespace DepartmentManagementAsp.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteEmployee([FromRoute] long id)
         {
+            Employee employee = new Employee { EmployeeId = id };
+            if (ValidateEmployeeOrBadRequest(employee, EmployeeValidator.ValidateOnDelete) is BadRequestObjectResult badRequest)
+            {
+                return badRequest;
+            }
+
             OperationResult operationResult =
-                await EmployeeRepository.DeleteEmployeeAsync(new Employee {EmployeeId = id});
+                await EmployeeRepository.DeleteEmployeeAsync(employee);
             return GetResultFromOperationResult(operationResult);
         }
 
         [HttpDelete("{employeeId}/from-department")]
         public async Task<IActionResult> RemoveFromDepartment([FromRoute] long employeeId)
         {
+            Employee employee = new Employee { EmployeeId = employeeId };
+            if (ValidateEmployeeOrBadRequest(employee, EmployeeValidator.ValidateOnDelete) is BadRequestObjectResult badRequest)
+            {
+                return badRequest;
+            }
             OperationResult operationResult = await EmployeeRepository.RemoveEmployeeFromDepartmentAsync(employeeId);
             return GetResultFromOperationResult(operationResult);
         }
@@ -62,6 +90,17 @@ namespace DepartmentManagementAsp.Controllers
             }
 
             return BadRequest(operationResult.Message);
+        }
+
+        private IActionResult ValidateEmployeeOrBadRequest(Employee employee, Func<Employee, IQueryable<Employee>, OperationResult> validator)
+        {
+            OperationResult validationResult = validator(employee, EmployeeRepository.GetEmployees());
+            if (!validationResult.Success)
+            {
+                return BadRequest(validationResult.Message);
+            }
+
+            return null;
         }
 
         private IEnumerable<Employee> BreakDepReferenceCycle(IQueryable<Employee> employees)
